🚚  Moves the telegram to a separate router
