+++ main.py
@@ -1,10 +1,10 @@
 import sentry_sdk
-from fastapi import BackgroundTasks, FastAPI, Request
+from fastapi import FastAPI
 from sentry_sdk.integrations.asgi import SentryAsgiMiddleware
 
 from config.base import config
 from routes import todoist as todoist_routes
-from services.telegram import auth as telegram_auth
+from routes import telegram as telegram_routes
 from utils import start_up
 
 
@@ -17,6 +17,7 @@ if config.sentry.dsn:
 
 app = FastAPI()
 app.include_router(todoist_routes.router)
+app.include_router(telegram_routes.router)
 app.on_event("startup")(start_up.startup_ensure_mongo_unique_id_indexes)
 app.on_event("startup")(start_up.startup_ensure_telegram_webhook)
 
@@ -31,16 +32,3 @@ app.include_router
 @app.get("/health")
 async def health():
     return {"message": "Hello World"}
-
-
-# Telegram Flow
-@app.post("/telegram/webhook")
-async def telegram_webhook(request: Request, background_tasks: BackgroundTasks):
-    data = await request.json()
-    text = data["message"]["text"]
-
-    if text.startswith("/start"):
-        config.logger.info("Received /start command from telegram")
-        background_tasks.add_task(telegram_auth.complete_telegram_verification, webhook_body=data)
-
-    return {"message": "Webhook received"}
